refactor: extract AnalyticsMapper, consolidate stock logic into AnalyticsService

+++ src/main/java/hei/school/library/dto/BookStockResponse.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 @Builder
 public class BookStockResponse {
   private UUID bookId;
-  private Integer total;
+  private long total;
   private BookCopyFormat byFormat;
 }
